スクリプトを改良

+++ Vertigo 2/localize.ps1
@@ -11,12 +11,6 @@ if ($isWin) {
     $jaPath = $jaPath -replace '/', '\'
 }
 
-if (-not (Test-Path -Path $dstPath)) {
-    Write-Warning "Error: 元の言語ファイルが見つかりません。Vertigo 2のインストールディレクトリに設置してください。";
-    Read-Host "Enterを押して終了";
-    exit 1
-}
-
 if (-not (Test-Path -Path $jaPath)) {
     Write-Warning "Error: 日本語の言語ファイルが見つかりません。導入手順に従ってください。";
     Read-Host "Enterを押して終了";
@@ -24,14 +18,31 @@ if (-not (Test-Path -Path $jaPath)) {
 }
 
 if (-not (Test-Path -Path $srcPath)) {
-    Copy-Item -Path $dstPath -Destination $srcPath
+    if (Test-Path -Path $dstPath) {
+        Copy-Item -Path $dstPath -Destination $srcPath
+    }
+}
+
+if (-not (Test-Path -Path $srcPath)) {
+    Write-Warning "Error: 元の言語ファイルが見つかりません。Vertigo 2のインストールディレクトリに設置してください。";
+    Read-Host "Enterを押して終了";
+    exit 1
 }
 
 $src = Import-CSV -Delimiter "," -Path $srcPath -Encoding UTF8
 $ja = Import-CSV -Delimiter "," -Path $jaPath -Encoding UTF8
 
-for ($i=0; $i -lt $src.Length; $i++) {
-    $src[$i].en = $ja[$i].ja
+$jaMap = @{}
+foreach ($row in $ja) {
+    $key = $($row.PSObject.Properties)[0].Value
+    $jaMap[$key] = $row.ja
+}
+
+foreach ($row in $src) {
+    $key = $($row.PSObject.Properties)[0].Value
+    if ($jaMap.ContainsKey($key)) {
+        $row.en = $jaMap[$key]
+    }
 }
 
 $src | Export-Csv -Delimiter "," -NoTypeInformation -Force -Path $dstPath -Encoding UTF8
